Moving CreateSearchResult logic into abstract class
