+++ src/Examine/LuceneEngine/LuceneSearchResultsBase.cs
@@ -1,0 +1,59 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Examine.LuceneEngine.Providers;
+using Lucene.Net.Documents;
+
+namespace Examine.LuceneEngine
+{
+    public abstract class LuceneSearchResultsBase : SearchResultsBase
+    {
+        /// <summary>
+        /// Creates the search result from a <see cref="Lucene.Net.Documents.Document"/>
+        /// </summary>
+        /// <param name="doc">The doc to convert.</param>
+        /// <param name="score">The score.</param>
+        /// <returns>A populated search result object</returns>
+        protected virtual ISearchResult CreateSearchResult(Document doc, float score)
+        {
+            var id = doc.Get("id");
+
+            if (string.IsNullOrEmpty(id) == true)
+            {
+                id = doc.Get(LuceneIndex.ItemIdFieldName);
+            }
+
+            var searchResult = new SearchResult(id, score, () =>
+            {
+                //we can use lucene to find out the fields which have been stored for this particular document
+                var fields = doc.GetFields();
+
+                var resultVals = new Dictionary<string, List<string>>();
+
+                foreach (var field in fields.Cast<Field>())
+                {
+                    var fieldName = field.Name;
+                    var values = doc.GetValues(fieldName);
+
+                    if (resultVals.TryGetValue(fieldName, out var resultFieldVals))
+                    {
+                        foreach (var value in values)
+                        {
+                            if (!resultFieldVals.Contains(value))
+                            {
+                                resultFieldVals.Add(value);
+                            }
+                        }
+                    }
+                    else
+                    {
+                        resultVals[fieldName] = values.ToList();
+                    }
+                }
+
+                return resultVals;
+            });
+
+            return searchResult;
+        }
+    }
+}
